Move matrix! into macros module, replace old one
